Add header file, and fix cpp file
