+++ src/codeFormatter.cpp
@@ -1,30 +1,9 @@
+#include "code_formatter.hpp"
+
 #include <iostream>
 #include <string>
 #include <algorithm>
 #include <cstdio>
-
-struct Section {
-  int delimiterIndex;
-  std::string frontCode;
-};
-
-struct DelimiterInfo {
-  char delimiter;
-  std::string normalFormat;
-  std::string onlyDelimFormat;
-  void (*taskFunc)(Section*, std::string&);
-};
-
-void inputCode(std::string& code);
-void formatCode(std::string& code);
-void indentCode(std::string& code);
-Section* getInputOneSection();
-bool remake(Section* section, std::string& result);
-int findDelimiter(const char input);
-void addIndentRow(int indentDepth, std::size_t workPoint, std::string& code);
-int getDiffIndent(const std::string& code, const std::size_t start, const std::size_t end);
-void taskNormal(Section* section, std::string& result);
-void taskSpace(Section* section, std::string& result);
 
 const int INDENT_WIDTH = 2;
 const DelimiterInfo delimiters[] = {
